int math, because float math is too slow

+++ src/include/libtemp/temp.h
@@ -1,6 +1,6 @@
 #ifndef LIBTEMP_TEMP_H
 #define LIBTEMP_TEMP_H
 
-float temp_sample();
+int temp_sample();
 
 #endif // LIBTEMP_TEMP_H
